requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.11
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.5
django-cleanup==5.2.0
django-on-heroku==1.1.2
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
fpdf==1.7.2
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
numpy==1.21.1
oauthlib==3.2.0
opencv-python==4.5.3.56
Pillow==8.3.1
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
